Modified `fixedToDescriptor` in runtimeC.c, trying to work around clang problem on Msys2.

+++ XCOM-I/runtimeC.c
@@ -1488,7 +1488,7 @@ putCHARACTER(uint32_t address, descriptor_t *str)
 
 descriptor_t *
 fixedToCharacter(int32_t i) {
-  return cToDescriptor(NULL, "%d", i);
+  return cToDescriptor(NULL, "%d", (int) i);
 }
 
 // Convert a BIT(1) through BIT(32) to FIXED.  As near as I can figure (see
